Use cacheds3storage
Use the package we created here, again:
https://github.com/ccnmtl/django-cacheds3storage

This should add back in integration between django-compressor and
django-storages. I realized this while reading django-compressor's docs
here:
https://django-compressor.readthedocs.io/en/stable/remote-storages/#using-staticfiles

This got lost during the django-storages / boto3 update, when I reverted
to django-storage's default classes for this. But, our cacheds3storage
is now compatible with the new boto3, so this should work.