Remove public-read global setting for ACL
This setting was originally introduced in this commit, because
apparently django-storages recommended it be set:

bfb7579f2679a7e2f2d0f6b356318d35a532a2f2

This is no longer the case - it's completely optional:
https://django-storages.readthedocs.io/en/latest/backends/amazon-S3.html

And when it's not set, it defaults to using the bucket's ACL, which is
what we want.

This change fixes our mediathread private bucket uploading issue - these
uploads should happen with a private acl, not 'public-read'.