Remove django-waffle from common settings
django-waffle is a feature flipper that we only rarely use these days.
Feature flags can be useful, but they also can introduce complexity to
an application.

In any case, I think it's fine to just manually include django-waffle in
the case that we do want to use a feature flag. It definitely doesn't need
to be included in all our projects.